Update event handler serialization
Allows users to specify a path to a value on a JS event object
they need. This is used when a use wants to track changes to an
input element, but needs to know the current 'value' of the input.
This is not normally possible because you can't serialize the whole
event object.

The specification change proposal can be found in this issue on vdom:

https://github.com/nteract/vdom/issues/89