Fix #19 (#20)
It is possible for events from the client to refer to elements
on the backend that have already been deleted. This is because
the frontend has not been sent the newly computed DOM model by
the time the event is produced - thus the target of the event
is stale.

In this case we now choose to ignore the event because the element
has been deleted.