Stop stale animation hooks on element update
Animation hook should be considered stale after an element
updates. This is because the local state known inside the hook's
lexical scope will not be the same used to render the next model.
Because these stale animation hooks will run just before the next
render they have a chance to check wether or not the element that
registered them is about to update and stop themselves.