Misc changes + 0.4.2 version bump (#56)
* Misc changes + version 0.4.2 bump

Fixes an issue with typing for element functions - protocol from
typing extensions does not produce correct behavior. Variable
positional and keyword arguments should be optional and not
required.

Reworked html to vdom function output. Before it ignored the
possibility of having multiple "root" elements - now, we wrap
whatever models are produced in a div and the children are the
roots of the html that was passed.

Allows multiple transforms to be passed for html to vdom and changes
how they are applied. Now they are run from roots to leafs instead
of the other way around. Feels easier to thing about that way.