Work to await futures
There's a number of places where we need to schedule coroutines
using `asyncio.ensure_future`. The problem with this is that
if the future this function returns is not awaited any errors
it encounters will not stop the program. They will just be logged.

These changes try to await those futures as well as possible, and
with the exception of a minor place within `FutureQueue`, do so.