Elements can render in thread executor (#68)
* Elements can render in thread executor

Adds a 'run_in_executor' keyword argument to the `element`
decorator that allows users to run a render function in a thread
executor. This is useful because certain long running operations
will prevent the next render from taking place and thus halt the
view until the operation completes. Processing these long running
updates in a thread executor unblocks the rendering loop.

In order to make this happen some underlying changes to the
`Layout` and communication protocol were made. In short, renders
are not run at the time `Layout.render` is called. They are
scheduled to be computed immediately and added to a queue when
they complete. The `Layout.render` function then pull from that
queue. The communication protocol was changed to include the
ID of the layout's root element because it may not be the first
element to get rendered and thus be placed in the queue.

Other minor changes were made - see `RenderBundle` type to see.

* Decouple animation from Layout

We could always have used `asyncio.ensure_future` to decouple
the `Element.animate` hook from the render loop in `Layout`.

* Work to await futures

There's a number of places where we need to schedule coroutines
using `asyncio.ensure_future`. The problem with this is that
if the future this function returns is not awaited any errors
it encounters will not stop the program. They will just be logged.

These changes try to await those futures as well as possible, and
with the exception of a minor place within `FutureQueue`, do so.