Dynamically Import React JS Libraries (#82)
Adds the ability to dynamically import React component libraries
and use them inside views. There's definitely a bit of work to do
here since not all React libraries seem to work out of the box.
Further, it should be ensured that the interface for imports makes
sense for VueJS too.

This adds an 'importSource' key to the root of the element spec
which contains the 'package' and a 'fallback'. However the bigger
change is that event handlers can accept multiple arguments. This
was brought about because certain callbacks for Ant Design send
more than one argument to their callbacks, thus event callback
data is now an array of arguments. So long as all those arguments
are JSON serializable everything should "just work".