refactor dispatchers
Simplify them greatly by making them normal functions.
This avoids a lot of indirection and complexity caused
by inheritance.

In addition this fixes some impropper usages of TaskGroups
which was causing problems in the SharedClientStateServer
implementations.

This refactor also allowed us to get rid of the clunky
HasAsyncResources util class. In the end it just added
confusion.