rework js module interface + fix docs
Because of changes to the JS interface, examples embedded in the
documentation had stopped working. Those issues have been fixed
now.

After resolving those problems though, a new issue cropped up in
which components from import sources were being unmounted on
every render. To fix that, the interface had to be changed.

There is no longer an exportsMount flag to indicate that the
module should be rendered in isolation. Now, if there exists
a render() and unmount() functions as named exports, components
from that module will be rendered in isolation.