fix 417 and 413
no more use of the id() function to generate unique
IDs for component and event handlers within the layout.

also includes a minor change to use a protocol (rather than
inheritance) for the component base-class. isinstance()
will not be a perfect check of type compatibility, but it
should be close enough