fix model mutation in client
The client-side model was being mutated on each update
in order to try and be more performant. This is a problem
though because React relies on checking the identity of
some parts of the model to see if things have changed.

No we create a copy of the model each time it changed