set target value onchange client-side (#694)
* set target value onchange client-side

This may not be a perfect solution. I worry that doing this may have unintended
consiquences. For example, what if someone wanted to have some sort of
auto-correct feature. Will setting the target value somehow overwrite
auto-corrections? From limited testing it seems to work fine, but that
testing was not extensive.

* changelog entry