Add <marquee> element
This commit simply adds the option to use the marquee element in reactpy. While yes, the marquee element has been deprecated, I think that it would be best to allow web developers the freedom to use this element even when it is deprecated. Also the marquee element is still usable in all major browsers.